@@ -99,14 +99,14 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.2.2",
-    "@angular/common": "~13.2.2",
-    "@angular/compiler": "~13.2.2",
-    "@angular/core": "~13.2.2",
-    "@angular/forms": "~13.2.2",
-    "@angular/platform-browser": "~13.2.2",
-    "@angular/platform-browser-dynamic": "~13.2.2",
-    "@angular/router": "~13.2.2",
+    "@angular/animations": "^14.3.0",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
     "@fortawesome/angular-fontawesome": "^0.10.1",
     "@fortawesome/fontawesome-svg-core": "~1.3.0",
     "@fortawesome/free-solid-svg-icons": "^6.0.0",
@@ -118,27 +118,27 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.2.3",
-    "@angular-eslint/builder": "13.1.0",
-    "@angular-eslint/eslint-plugin": "13.1.0",
-    "@angular-eslint/eslint-plugin-template": "13.1.0",
-    "@angular-eslint/schematics": "13.1.0",
-    "@angular-eslint/template-parser": "13.1.0",
-    "@angular/cli": "~13.2.3",
-    "@angular/compiler-cli": "~13.2.2",
+    "@angular-devkit/build-angular": "^14.2.13",
+    "@angular-eslint/builder": "14.4.0",
+    "@angular-eslint/eslint-plugin": "14.4.0",
+    "@angular-eslint/eslint-plugin-template": "14.4.0",
+    "@angular-eslint/schematics": "14.4.0",
+    "@angular-eslint/template-parser": "14.4.0",
+    "@angular/cli": "^14.2.13",
+    "@angular/compiler-cli": "^14.3.0",
     "@types/jest": "^27.4.0",
     "@types/node": "^17.0.18",
     "@types/semver": "^7.3.9",
     "@types/vscode": "^1.64.0",
     "@types/xml2js": "^0.4.9",
-    "@typescript-eslint/eslint-plugin": "5.12.0",
-    "@typescript-eslint/parser": "5.12.0",
-    "eslint": "^8.2.0",
+    "@typescript-eslint/eslint-plugin": "^5.36.2",
+    "@typescript-eslint/parser": "^5.36.2",
+    "eslint": "^8.23.0",
     "jest": "^27.5.1",
     "jest-preset-angular": "^11.1.1",
     "npm-watch": "^0.11.0",
     "ts-loader": "^9.2.6",
-    "typescript": "~4.5.2",
+    "typescript": "~4.8.4",
     "webpack": "^5.69.1",
     "webpack-cli": "^4.9.2"
   },
